민형이형 리액트뷰 / 우석이 고객센터 시세조회 / 상빈이형 마이페이지 유저샵 트레이드

+++ src/main/java/com/Reboot/Minty/categories/entity/TopCategory.java
@@ -1,10 +1,14 @@
 package com.Reboot.Minty.categories.entity;
 
+import com.Reboot.Minty.tradeBoard.entity.TradeBoard;
 import jakarta.persistence.*;
 import lombok.Getter;
 import lombok.Setter;
 import lombok.ToString;
 import org.springframework.transaction.annotation.Transactional;
+
+import java.util.ArrayList;
+import java.util.List;
 
 @Entity
 @Table(name="top_categories")
@@ -15,8 +19,8 @@ import org.springframework.transaction.annotation.Transactional;
 public class TopCategory {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private int id;
-
+    private Long id;
     private String name;
+
 
 }
